+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheet.swift
@@ -622,13 +622,10 @@ struct VBottomSheet_Previews: PreviewProvider {
                             ScrollView(content: {
                                 VStack(spacing: 0, content: {
                                     ForEach(0..<20, content: { number in
-                                        VListRow(
-                                            uiModel: .noFirstAndLastSeparators(isFirst: number == 0),
-                                            content: {
-                                                Text(String(number))
-                                                    .frame(maxWidth: .infinity, alignment: .leading)
-                                            }
-                                        )
+                                        Text(String(number))
+                                            .frame(maxWidth: .infinity, alignment: .leading)
+                                            .padding(.horizontal, 15)
+                                            .padding(.vertical, 9)
                                     })
                                 })
                             })